+++ HISTORY.rst
@@ -4,7 +4,7 @@ Changelog
 2.2.2 (unreleased)
 ------------------
 
-- Add Mamoto support (CastleCMS API control panel settings, Twitter share counting and Analytics display);
+- Add Mamoto support (CastleCMS API control panel settings, Twitter & Facebook share counting and Analytics display);
   remove EOL'd Facebook and LinkedIn API social counts
   [tkimnguyen]