+++ HISTORY.rst
@@ -4,7 +4,9 @@ Changelog
 2.2.3 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Add Mamoto support (CastleCMS API control panel settings, Twitter & Facebook share counting and Analytics display);
+  remove EOL'd Facebook and LinkedIn API social counts
+  [tkimnguyen]
 
 
 2.2.2 (2018-11-26)