+++ HISTORY.rst
@@ -4,6 +4,9 @@ Changelog
 2.2.3 (unreleased)
 ------------------
 
+- Handle incorrectly configured backend url/sheild settings
+  [vangheem]
+
 - Add Mamoto support (CastleCMS API control panel settings, Twitter & Facebook share counting and Analytics display);
   remove EOL'd Facebook and LinkedIn API social counts
   [tkimnguyen]